refactor: rename CustomSnackBar package to SnackBar
